+++ dartagnan/src/main/java/com/dat3m/dartagnan/expression/Atom.java
@@ -2,7 +2,7 @@ package com.dat3m.dartagnan.expression;
 
 import com.dat3m.dartagnan.expression.op.COpBin;
 import com.dat3m.dartagnan.expression.processing.ExpressionVisitor;
-import com.dat3m.dartagnan.parsers.program.utils.TypeException;
+import com.dat3m.dartagnan.parsers.program.exception.TypeException;
 import com.dat3m.dartagnan.program.Register;
 import com.dat3m.dartagnan.program.event.Event;
 import com.dat3m.dartagnan.program.memory.Location;
