+++ dartagnan/src/main/java/com/dat3m/dartagnan/expression/IfExpr.java
@@ -1,7 +1,7 @@
 package com.dat3m.dartagnan.expression;
 
 import com.dat3m.dartagnan.expression.processing.ExpressionVisitor;
-import com.dat3m.dartagnan.parsers.program.utils.TypeException;
+import com.dat3m.dartagnan.parsers.program.exception.TypeException;
 import com.dat3m.dartagnan.program.Register;
 import com.dat3m.dartagnan.program.event.Event;
 import com.dat3m.dartagnan.program.memory.Location;
