+++ dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/ParserLitmusPPC.java
@@ -2,7 +2,7 @@ package com.dat3m.dartagnan.parsers.program;
 
 import com.dat3m.dartagnan.parsers.LitmusPPCLexer;
 import com.dat3m.dartagnan.parsers.LitmusPPCParser;
-import com.dat3m.dartagnan.parsers.program.utils.ParserErrorListener;
+import com.dat3m.dartagnan.parsers.program.exception.ParserErrorListener;
 import com.dat3m.dartagnan.parsers.program.utils.ProgramBuilder;
 import com.dat3m.dartagnan.parsers.program.visitors.VisitorLitmusPPC;
 import com.dat3m.dartagnan.program.Program;
