+++ dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/ParserLitmusX86.java
@@ -2,7 +2,7 @@ package com.dat3m.dartagnan.parsers.program;
 
 import com.dat3m.dartagnan.parsers.LitmusX86Lexer;
 import com.dat3m.dartagnan.parsers.LitmusX86Parser;
-import com.dat3m.dartagnan.parsers.program.utils.ParserErrorListener;
+import com.dat3m.dartagnan.parsers.program.exception.ParserErrorListener;
 import com.dat3m.dartagnan.parsers.program.utils.ProgramBuilder;
 import com.dat3m.dartagnan.parsers.program.visitors.VisitorLitmusX86;
 import com.dat3m.dartagnan.program.Program;
