+++ dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/ProgramParser.java
@@ -1,6 +1,6 @@
 package com.dat3m.dartagnan.parsers.program;
 
-import com.dat3m.dartagnan.parsers.program.utils.ParsingException;
+import com.dat3m.dartagnan.parsers.program.exception.ParsingException;
 import com.dat3m.dartagnan.program.Program;
 import org.antlr.v4.runtime.CharStream;
 import org.antlr.v4.runtime.CharStreams;
