+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/llvm/utils/LlvmFunctions.java
@@ -6,7 +6,7 @@ import com.dat3m.dartagnan.expression.IConst;
 import com.dat3m.dartagnan.expression.IExprBin;
 import com.dat3m.dartagnan.expression.op.BOpUn;
 import com.dat3m.dartagnan.expression.op.IOpBin;
-import com.dat3m.dartagnan.parsers.program.utils.ParsingException;
+import com.dat3m.dartagnan.parsers.program.exception.ParsingException;
 
 import java.util.Arrays;
 import java.util.List;
