+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/llvm/utils/SmackPredicates.java
@@ -13,7 +13,7 @@ import com.dat3m.dartagnan.expression.ExprInterface;
 import com.dat3m.dartagnan.expression.IConst;
 import com.dat3m.dartagnan.expression.IExprBin;
 import com.dat3m.dartagnan.expression.IfExpr;
-import com.dat3m.dartagnan.parsers.program.utils.ParsingException;
+import com.dat3m.dartagnan.parsers.program.exception.ParsingException;
 
 public class SmackPredicates {
