+++ dartagnan/src/test/java/com/dat3m/dartagnan/parsers/program/utils/ProgramBuilderTest.java
@@ -3,6 +3,7 @@ package com.dat3m.dartagnan.parsers.program.utils;
 import com.dat3m.dartagnan.parsers.program.ProgramParser;
 import com.dat3m.dartagnan.utils.ResourceHelper;
 import org.junit.Test;
+import com.dat3m.dartagnan.parsers.program.exception.ParsingException;
 
 import java.io.File;
 import java.io.IOException;
