+++ svcomp/src/main/java/com/dat3m/svcomp/utils/SVCOMPSanitizer.java
@@ -9,7 +9,7 @@ import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
 
-import com.dat3m.dartagnan.parsers.program.utils.ParsingException;
+import com.dat3m.dartagnan.parsers.program.exception.ParsingException;
 
 public class SVCOMPSanitizer {
